cpu: jit_utils: move jit_utils to src/cpu
