+++ src/cpu/aarch64/jit_generator.hpp
@@ -26,13 +26,15 @@
 
 #include "cpu/aarch64/cpu_isa_traits.hpp"
 
-#include "cpu/aarch64/jit_utils/jit_utils.hpp"
+#include "cpu/jit_utils/jit_utils.hpp"
 
 #define STRUCT_ALIGN(al, ...) __VA_ARGS__ __attribute__((__aligned__(al)))
 
 #define DECLARE_CPU_JIT_AUX_FUNCTIONS(jit_name) \
     const char *name() const override { return STRINGIFY(jit_name); } \
     const char *source_file() const override { return __FILE__; }
+
+static const size_t CSIZE = sizeof(uint32_t);
 
 namespace dnnl {
 namespace impl {
@@ -371,7 +373,7 @@ private:
         if (!is_initialized()) return nullptr;
         const uint8_t *code
                 = reinterpret_cast<const uint8_t *>(CodeGenerator::getCode());
-        register_jit_code(code, getSize());
+        register_jit_code(code, getSize() * CSIZE);
         return code;
     }
