+++ src/cpu/jit_utils/linux_perf/linux_perf.hpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2019-2020 Intel Corporation
+* Copyright 2019-2021 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -14,23 +14,26 @@
 * limitations under the License.
 *******************************************************************************/
 
-#ifndef CPU_X64_JIT_UTILS_JIT_UTILS_HPP
-#define CPU_X64_JIT_UTILS_JIT_UTILS_HPP
+#ifndef CPU_JIT_UTILS_LINUX_PERF_LINUX_PERF_HPP
+#define CPU_JIT_UTILS_LINUX_PERF_LINUX_PERF_HPP
 
-#include <cstdlib>
+#ifdef __linux__
+#include <cstddef>
 
 namespace dnnl {
 namespace impl {
 namespace cpu {
-namespace x64 {
 namespace jit_utils {
 
-void register_jit_code(const void *code, size_t code_size,
-        const char *code_name, const char *source_file_name);
+void linux_perf_jitdump_record_code_load(
+        const void *code, size_t code_size, const char *code_name);
 
-}
-} // namespace x64
+void linux_perf_perfmap_record_code_load(
+        const void *code, size_t code_size, const char *code_name);
+} // namespace jit_utils
 } // namespace cpu
 } // namespace impl
 } // namespace dnnl
+#endif
+
 #endif
